move the update_status timer removal a bit earlier
seems this was a bit too late, it causes some kind of conflict
with the `upload_status(1)` calls I think, and results in
workers dying with `recursive api_call is fatal` errors. So
move it back to just after the file uploads are done, but
before all of those calls.