Give uploading jobs more time before considered dead
For this I reworked the dead worker detection not to work on timers, but
query every couple of minutes (I went from 20 minutes to 2 minutes not
to make this such an unlikely event) if the worker was updated in time.
And the timeout depends on the state of the job. 10 seconds for running,
1000 for uploading