Add new job state (#848)
* Avoid dead job detection for jobs uploading large files

When the job finishes, we set the job state to 'uploading' so the webui
can differ between those jobs and jobs that are supposed to update its
status every couple of seconds. This way we can have different timeouts
for these

* Give uploading jobs more time before considered dead

For this I reworked the dead worker detection not to work on timers, but
query every couple of minutes (I went from 20 minutes to 2 minutes not
to make this such an unlikely event) if the worker was updated in time.
And the timeout depends on the state of the job. 10 seconds for running,
1000 for uploading

* Prevent timeouts during uploading of files

by disabling keep alive for our webapi user agent

* Adopt the test case for dead worker detection

There is no timer trickery required anymore

* Send KILL to worker process group when not reacting to TERM

* Use a timeout for all jobs in execution

Make UPLOADING the exception. We used to ignore WAITING jobs for
dead workers, but this sounds unfair - these workers can die too