dead worker check: use 15 secs not 10, better names, comments (#864)
As discussed in #848, it's kind of racy to have the workers
POST a status update every 10 seconds, and have the dead worker
check consider any worker which hasn't updated in the last 10
seconds 'dead'. It doesn't allow for there being a slight delay
in POSTing the update (which can happen if the worker is busy
doing something else when the 10 second timer should trigger
the update) or in the update being actually received and
processed by the server. So let's just change the threshold for
the check to 15 seconds.

While I'm at it, I figured we could make the subroutine names
a bit clearer: since the DB query subroutine isn't really the
ultimate arbiter of whether the worker is truly 'dead' or not
any more, let's call that `_get_stale_worker_jobs` instead of
`_get_dead_worker_jobs` (and call its result `$stale_jobs` not
`$dead_jobs`). This should make the code a bit less confusing
to anyone reading it fresh.

Also, coolo's changes from #848 changed the behaviour here
without updating all the comments; update a couple of comments
and explain a bit more in them.