comment the home detection for scheduler/websockets better
I think this deserves a fairly verbose comment because it's
pretty obscure; I can see us coming back to it in a few months
and wondering why the hell the 'require' is there. Also, if you
break this it's not really incredibly obvious, so someone could
take it out, not see that anything explodes, and think it must
be useless. So let's explain clearly what's going on.