turn job_notify_workers into a websockets IPC call
As noted by @aaannz in #883, the current job_notify_workers has
become a bit weird. You have to either import the scheduler or
send it an IPC call in order to use it, but all the scheduler
does is turn around and send an IPC call to websockets. We often
need to notify workers from places other than the scheduler, so
having it there really doesn't make a lot of sense any more. So
instead, let's just make it a websockets IPC call directly.