Try harder to avoid long running pre-processing
If many workers grab a job at the same time, they all go through the
scheduler bottleneck, wo will assign every worker a job one by one.
But the last couple of job_grab requests will actually get a timeout
and the worker will believe it didn't get a job.

if it then later grabs another job, we need to do something with the
previously assigned job. Just saying no is a problem as both the worker
keeps being unused and the job stays in running (pre-processing).

So now I set the previous job to incomplete and duplicate it. Beside
that I also made sure that cancelling a job kills the worker reference,
so it can grab another one.