Avoid DBUS dead locks
Instead of calling blocking DBUS methods into websockets when notifying the worker,
we do a non blocking call into the scheduler, which then emits a signal.

This signal is then listened from websockets (whenever it has time :) to notify
the workers. None of that blocks and as such can't deadlock