Add 'fixed' asset subdirectories for fixed assets (#928) (#945)
This adjusts all the places that look for asset files to check
both in the current location and in a new possible location,
which is in a 'fixed/' subdirectory of the normal asset folder
for that asset type. It also adjusts the `limit_assets` gru
task to ignore files in the 'fixed' subdirectories. This
replaces the current special-casing of disk image assets created
by user '0'. That approach has become problematic now both SUSE
and Fedora are testing disk images that come from the distro
composes; these disk images are never getting cleaned up and
that causes disk space exhaustion. This also allows for 'fixed'
ISO and other assets, which the current approach does not.

As part of this change, we do a bit of refactoring to avoid
duplicating the code for constructing the possible file paths,
and have the worker code use the asset path definitions from
Utils rather than duplicating them.