Fix group overview page when job group has no description
The return value of `rendered_description` is passed to the
group_overview template as `$description`. If there is no
description text for the group, it winds up as undefined, and
that causes Mojo to barf when rendering the template and you
get an error when you visit the group overview page. So let's
have `rendered_description` return the empty string rather than
nothing at all when there's no description text, and now the
template behaves as intended.