Don't free worker before artefacts are uploaded (POO #14378)
0b78fc5f broke artefact uploading on job cancellation. Because
the job `cancel()` method now frees the worker (so far as the
schema is concerned) immediately after telling the worker to
cancel the job, by the time the worker process reaches the
artefact upload bits of `_stop_job()`, the schema job instance
has no worker, and when the worker process tries to upload the
artefacts, the schema job instance `create_artefact()` method
throws a 404 because so far as it knows, it no longer has a
worker.

The change was intended to mitigate a problem caused by the
dbus deadlock issue, where the server never actually handled
the `set_done` API call the worker process sends at the end
of `_stop_job`, so the worker was never freed - but as the dbus
deadlock issue has now been addressed, we shouldn't need to
worry about that any more.