Rejig compressed asset download, don't leave compressed files (#955)
There's two problems with the current compressed asset download
code:

1) The compressed files will never be cleaned up
2) Specifying FOO and FOO_DECOMPRESS_URL will not work

1) is because the compressed file is download to the asset
directory then uncompressed and left there; because the
compressed file itself will never be a job asset, the asset
cleanup code will never clean it up.

2) is because the current compressed asset download code always
uncompresses the compressed file to the same file name with the
final extension removed, even if you passed in a completely
different file name as the FOO parameter. In this situation,
we should uncompress the downloaded data to the name given as
FOO in the appropriate path.

This should fix both problems. Instead of passing the path for
the compressed file from `schedule_iso` to `download_asset` and
then having `download_asset` decide the final uncompressed file
name, we always pass the ultimately desired asset path, the
download URL, and the uncompress flag from `schedule_iso` to
`download_asset`. If decompression is requested, the downloaded
data is moved into a temporary file first and then uncompressed
to the requested final asset path and the temp file removed. If
not, the downloaded data is directly moved to the requested
final asset path.

We also extend the `schedule_iso` testing a bit to cover these
situations (and do stricter checking of some existing ones).