Fix worker error reporting when asset cannot be found (#963)
I introduced a problem in baac24bc by using `locate_asset` here.
Previously when the asset wasn't found, the variable (`$iso` or
whatever) would be the path where we were looking for it; now,
it's undef. So the error message won't tell us what asset it was
looking for, and additionally we get a (non-fatal) perl error
constructing the error message. So let's keep the asset filename
around for the purpose of creating the error message, and tweak
the wording a bit.