Add support for github 'bugrefs'
Add support for generic github repo markers as 'bugrefs
Using the generic profile
'<marker>[#<project/repo>]#<id>'
for all bugrefs, especially for github, allows to
track all issues reported on github and especially
test fixing pull requests to be recorded as 'bugrefs'
with the same icon as for "test issues", i.e. poo#.
This way there is no need to create a ticket just
to track a test issue for which already a github
pull request exist because otherwise the URL to the
github PR would not mean anything special to github
and therefore not mark a job as "labeled".

Also, pasting full github URLs is supported, same as
for all other bugrefs, meaning that the URL is
abbreviated into short form bugrefs.

The pattern follows the openSUSE/OBS recommendation.