API jobs list: add params 'before' and 'after' (#996)
I want to do something like 'find the job before job X that has
the same scenario' - this is something we do internally in a
few places, e.g. the 'previous results' page in the web UI. But
it's very difficult to do via the API, because you really can't
say "I only want jobs with IDs lower than X".

It's easy enough to add, though! So add 'before' and 'after'
params for the API method and also for the `complex_query`
method which backs it. While I'm at it, add `page` to the args
for the API method too (it's pretty useless to support `limit`
without `page`, and `complex_query` already handles it) and
cover all these bits in the tests, plus test the variant forms
for specifying multiple `ids` args (since I wrote that a while
back but it's not tested yet).