Add issue reporting links to needle preview screen
Using the available context of test results a URL for
corresponding issue tracker can be generated. The
fields in the to be opened issue can already be
prefilled. This can be extended later on even more.

The main customization of which issue trackers and
with what content is done in a branding template as
valid for openSUSE/SUSE. The mapping to bugzilla
products for both openSUSE products as well as the
SLE family is also done in there.