Add test database customization hook
For now we used two ways to setup a test database, either an empty one with
custom additions of entries tailored for a test or based on our fixtures which
have a good common set. Tests based on phantomjs so far always relied on the
database fixtures. The problem with that is by adding more entries to the
fixtures many tests need to be adapted, conflicting corner cases can not be
represented very well, the fixtures maintenance effort increases and the tests
become slower.

This commit provides a way to pass a custom hook for database creation
allowing a customization of database entries as necessary for the specific
tests, be it based on a clean database where one just wants to add what is
necessary or creating a database based on the fixtures and tailoring it for
the test in question.

Documentation for the feature is in t/ui/PhantomTest.pm and an example is in
t/ui/01-list.t.