Restart the mojo worker every 2000 requests (#1046)
To avoid the worker growing memory usage we restart it after 200*10
requests and give it a lot of time to finish requests. Mojolocious 7.10
has a bug, that during restart it will still accept more requests from
apache, so we need to limit the number of requests accepted (-r)
and give the restart enough time to finish all of them (-G).

This bug caused several sudden deaths in uploads, e.g. see
https://progress.opensuse.org/issues/13876