Use proper die message on duplicate job name
The 'NAME' field is checked to be unique for jobs and can fail but the value
'0' is not a good choice for that cause. Triggering a single job was plain
wrong assuming the job return value to always be a valid argument which was
then catched and forwarded to the API caller in a totally misleading error
message. For the iso the job creation failure was just ignored in before. Now
the code would just die on the iso post request which should be much more
helpful.