Fix various issues with fedmsg comment events (#1055)
Now I got around to checking on the comment events that were
added to the Fedmsg plugin, there are several problems. This
should fix most of them.

The modname for the comment events was entirely wrong, because
the modifications done for job events weren't being done for
comment events, so we got a modname with the useless 'openqa'
on the front and still with underscore separators instead of
dots. As a bonus, we now replace *all* underscores with dots,
not just the first (so `job_update_result` was actually also
broken before, I just didn't notice because we never much used
it).

We would never actually emit a message for comment deletion,
because `on_comment_event` tries to find the comment and bails
if it can't; obviously it won't find a comment that was just
deleted. For now, just send out a very bare message in this
case that includes only the comment ID; we could make this
better, I think, but it would make more sense as a separate
commit.

`on_comment_event` also lied about what its args were. The
first arg is the user ID, not the comment ID.

No tests were included in the initial commit. If it had had any,
these errors would've been much more obvious. I wrote some.