Add retry for failed uploads on server errors (5xx) (#1053)
Add a check to ensure that if the upload failed due to server being
unavailable (status 5xx) that will be executed up to $retry_limit times
for $tics of 1 second each.

The fix was tested by inverting the condition (i.e. checking for code
200), and adding a finished event hook (To ensure that the upload was
being issued properly) due to the inhability to create a proper test.