Comment events: Drop unneeded fedmsg conditional, include id
This touches both Fedmsg and AMQP comment event emitters. We
no longer need the conditional in the Fedmsg plugin to produce
a different hash in the case of comment deletion, as the API
code was changed to emit the mojo event *before* rather than
after deleting the comment. Also include the comment ID in the
emitted events, since it doesn't hurt anything and could be
useful (seems odd to emit an event with lots of info on the
subject of the event, but not its actual ID).