Sort needles by RMSE instead of MIN
So far we sorted the needles while displaying them by the minimum of the
similiarity of the match areas. The problem is that this makes needles
more equal than they are. In case if you have 2 needles that both mismatch

Needle A:
  Area 1 matches 0%

Needle B:
  Area 1 matches 100%
  Area 2 matches 0%

Both needles have a minimum of 0% and so they would display as:

0%: A
0%: B

With this change, we display the needles by their root mean square error,
which results in:
71%: B
0%: A