Have initdb exit 4 when no init or upgrade is done
022553fd made `initdb --init_database` silently return 0,
instead of crashing with an uncaught exception, when the db
is already initialized. The exception was bad, of course, but
the new behaviour is a bit of a problem for things like ansible
or puppet or salt, where it's Good Practice for actions to be
aware of whether they changed anything. It's not longer possible
to know whether a run of `initdb` a) initialized the database, b)
upgraded the database or c) did nothing because the database was
fine.

So, this makes it print an informative message about what it did
and exit 0 if it actually did anything; if the db was already
initialized and up to date, it exits 4 (a distinctive code that
ansible plays etc. can recognize).