Use locking for db check, run check on most db connections (#1095)
This is an attempt to address POO #15540. There are two major
changes.

First, we `flock` the db config file (chosen as we must be
able to read it, and the code to find it was already present)
to try and ensure that only one openQA process (one of the
server processes, gru, or any of the admin scripts that handle
the database) can run the deployment check at a time; this is
to ensure multiple processes don't try to deploy or upgrade the
db schema at once.

Second, we have the first call to `connect_db` for any given
process run the deployment check (because `connect_db` caches
the schema, subsequent calls won't run the check). This should
ensure that the db is always checked, and deployed or upgraded
if needed, when *any* openQA server process tries to use it.
Previously, only the webui process ran the deployment check,
which wasn't really sufficient.

We allow passing `check => 0` to `connect_db` to disable this,
so `initdb`, `upgradedb`, and some tests which actually want
to get a schema without doing the deployment check can do so.