Improve database deployment locking
On closer inspection of the `flock` documentation, I find that
we must open the lock file with write intent to make sure we
can get an exclusive lock (`LOCK_EX`); it *may* work if the
file is only open with read intent, but only if a native flock
function is being used (rather than emulation via lockf or
fcntl).

By default `database.ini` is installed with 640 permissions,
and both Fedora and openSUSE packages preserve this, so I
think it's OK to expect that in a typical deployment, we will
be able to write it. But there's no real *need* for this, it
*could* be made read-only by a paranoid admin.

So to do our best, we try to open the file in append mode, and
if that doesn't work, we open it in read mode. If we can't open
it in either mode, or if we can't get an exclusive lock after
opening it, we die.