Test both phantomjs and chrome on travis
You can set SELENIUM_CHROME to test with chrome, but phantomjs is
still the default. There are some special cases where you have to
take care about the difference though.

The cursor handling in input fields is a bit different in chrome
and alerts are handled properly while we need workarounds for
phantomjs.

Testing with Chrome has the advantage that you can see the browser
that is testing - and possibly spot problems while the tests fly by.
And if you add a sleep in the test during development, you can actually
inspect the HTML of the test data.

If you want to make SELENIUM_CHROME for you, it's advisable though
to run Xvfb :99 & and then run DISPLAY=:99 prove

I took the chance to change some test APIs