Don't use config file as database lock file (#1157)
Another tweak to the database deployment locking stuff. As
discussed in #1095, using the db config file for locking is not
a great idea; it requires the server be able to write to the
config file, which is denied by the AppArmor config on SUSE,
and that's probably a sensible precaution (it's never wise to
let webapps write to more stuff than is strictly needed).

So here's yet another try: let's rely on the fact that prjdir/db
(usually /var/lib/openqa/db) is made writeable by the server in
the install process and the SUSE/Fedora packages, and put the
lock file there, just assuming we are able to write to that
location.

We also adjust the full-stack dir layout to be more like a real
openQA install (the `db` directory under the `openqa` directory),
otherwise that test fails because it can't write to that non-
existent location.