Avoid OpenQA::IPC::ipc being called premetaure
This singleton is very dangerous as if you call it wrong the first time,
it will stick to that. And that's what happened in the full-stack test.
The websocket server was calling the scheduler with ipc->scheduler during
its startup phase - and this set the singleton to its own event loop.
Just that we never called that event loop - and so the dbus service was
stuck on the bus and caused everything else to lockup.

I overreacted a bit and now expect the fake tests to set an environment
variable if they want the testing event loop and the websocket service
to always call ipc with 1 - which will use the mojo loop