Fix build lookup for single groups (poo#15988)
Fixes a problem for '/tests/overview' when no build is specified, for example
to show the most recent build. Problem was that an 'unrelated' build could
show up as the groupid was not forwarded to the database query.

Also introduces a check for multiple groups and no build which is currently
not supported.

Probably introduced by dce250d.

Related progress issue: https://progress.opensuse.org/issues/15988