Add support for multiple distri/version/group on /tests/overview
This adds support to specify zero to multiple of distri/version/group
parameters to show multiple ones as well as look up corresponding entries from
the job results table. Every group of results will be displayed just as before
as a table on the tests overview page. Previously the title for each table was
just the "flavor" where now the "distri" as well as "version" is also
specified. For the special case of having just one "distri" and/or "version"
it is displayed just as before, i.e. only the "flavor" is shown. This also
works if "distri" and/or "version" have not been specified as query parameters
but the search in the job results based on the other parameters resolved to
just one value of these.

With this change also no query parameter for the /tests/overview route is
mandatory anymore.

Also extracting a partial template for the result table itself.

Related progress issue: https://progress.opensuse.org/issues/13712