Allow override of the usual setting precedence order (#1200)
As suggested by @aaannz in #684, allow specifying a variable
as `+VARIABLE` to signify that it should not be overridden by
another setting of `VARIABLE` later in the usual precedence
order. For instance, usually if the same variable is set in both
a product and a test suite, the test suite setting wins; with
this change, if the product specified `+VARIABLE` and the test
suite `VARIABLE`, the product setting would win. One major use
for this could be to prevent an `HDD_1` set from an ISO post
request overriding the value set in templates for a test which
uses a disk image uploaded by another test.

If multiple things specify `+VARIABLE`, I believe the one that
comes highest in the normal precedence order should win out. We
could, of course, extend this to give `++VARIABLE` even greater
precedence, and so on...:)