Fix UI test skipping when modules not available (#1211)
Also don't require Chrome when testing with PhantomJS, and vice
versa. @coolo broke this pretty bad recently; please be careful
not to just throw `use` around in this stuff, it breaks the
case where Selenium is unavailable entirely.

I also took the chance to clean up the implementation and the
message a bit.