Allow configuration of how builds are sorted in web UI
Currently, openQA assumes that `BUILD` values can be sorted in
'version number' style, when sorting the lists of builds for the
web UI index page and job group overview pages. SUSE follows
this convention, but Fedora doesn't and never has. Our compose
`BUILD` values are productmd 'compose IDs', like this:

Fedora-26-20170301.n.0
Fedora-Rawhide-20170301.n.0

and we have recently started testing updates, whose `BUILD`
values look like this:

FEDORA-2017-561942797a

where the final string is just a randomly-generated identifier,
it cannot be relied upon to sort in any way at all. It's not
really plausible to change these `BUILD` values into ones that
can be sorted like version numbers, so instead, let's make it
possible to specify at the job group level whether `BUILD`
values for jobs in that group can be sorted as versions or not.
If not, `compute_build_results` will stick with the initial
sort order, which is to use the time a job was most recently
created for each build as the key.

Note on the removal from `deploy.t`: the creation of a system
user during database schema upgrade only happens on upgrade
from version 50 or earlier. The test works by deploying with
schema version (current minus 2). Since the 'current' schema
version is now 53, the test deploys schema version 51, and the
upgrade process will not create the 'system' user. I think it's
reasonable to drop the test at this point as we've proven that
the 50->51 upgrade code works, and it's never going to change
now. That's the only thing the test was testing.