Refactor to allow CACHEDIRECTORY to replace share directory
Changes applied in this commit will alow the openQA to replace
share directory settings comming from the webUI with the CACHEDIRECTORY
value combined with the host, so multiple webUIs are supported now.

Assets still need to get a random filename, but they're at least separately
stored in the db as per webUI now.