Merge 'Status' and 'Connected' columns in workers overview
Status column was mistakenly printing Dead even for live workers
due to quering database instead of websocket connection as
Connected status was.
This PR merge information of both into the Status. When worker
is idle, websockets connection is queried and Online/Offline is
displayed, else information about job is available.