Do not upload empty serial_terminal.txt
Generally we enable the virtio terminal through the Machine settings which
means it is enabled for many tests which don't use it. QEMU creates a log file
for the virtio serial device whether it is used or not, so we have a lot of
empty log files. This was reported to be confusing so this commit only
uploads the file if it has data in it.

The behavior for other files is deliberately left unchanged so as to avoid
hiding the existence of an empty file when it might be used to diagnose a
problem.

Reported-by: Maxlin