Relax the regex for build values in tag: comments again
This regex was still too restrictive, even after the last time
it was loosened. It rejects BUILD values with a period after
the final dash, which all Fedora BUILDs have, because they look
like this:

Fedora-25-20170327.n.0

I don't see any value to this restriction, so I think we can
just simplify the regex to allow any string consisting only of
dashes, periods, @s, digits and word characters.

This was preventing tagging builds as 'important' working for
Fedora.