Don't obsolete jobs for VERSION-BUILD tagged builds
https://github.com/os-autoinst/openQA/pull/1165 allowed builds
to be tagged by VERSION-BUILD as well as just BUILD. That is,
if the BUILD is 1234 and the VERSION is 99, you can tag it like
this:

tag:1234:important:description

Or like this:

tag:99-1234:important:description

The code which shows the tag next to the build in the web UI
handles this okay. However, no-one adjusted the job method
`cancel_by_settings` - which decides which existing jobs to
cancel when some new jobs have been created - to handle this
form of tag, and so if you'd tagged a build using this form of
tag, then its jobs can still be 'obsoleted' in this way. This
fixes the problem by excluding jobs from the list of cancel
candidate if either the job's BUILD or 'VERSION-BUILD' matches
one of the entries in the 'important' list (instead of only
checking the job's BUILD).