Replace old 'important' behaviour with 'ignore_failure'
This partly reverts commit
487fcdbd56f5e46bd569a5ffcbbbb0e90da5744a, which made openQA's
result calculation ignore the 'important' test flag entirely.

This is really inconvenient to Fedora, we found the ability to
have 'unimportant' tests quite handy and have used it in a few
different ways. Adapting to its removal has proved very hard.
So this partly restores the 'important' behaviour, but with an
'ignore_failure' flag instead of 'important'. This both inverts
the default from before, and behaves slightly differently: no
soft failure mode is involved. 'ignore_failure' does exactly
what it sounds like - a failure or soft failure of the module
will never result in failure of the overall test. If a test is
made up entirely of 'ignore_failure' modules, it will always
have 'PASSED' as its overall result. This behaviour is fine for
the cases Fedora cares about, while still allowing the removal
of the logic to handle different soft failure cases (#1293).

There will be a corresponding PR for os-autoinst, but as it
doesn't actually use the flag directly, the changes there are
quite small.