Makefile: Allow to skip checkstyle
As the 'checkstyle' test can take some time provide a switch to skip it on
demand which can be useful when for example one wants to collect coverage from
a single test and execute this in a faster manner.

Example use: `make CHECKSTYLE=0 PROVE_ARGS=t/full-stack.t coverage-html` to
execute the full stack test and collect coverage without needing to conduct
the long-running style check every time.