Refactor Cache.pm to use SQLite instead of json
This changes in the Cache.pm attempt to solve the race condition where
more than two workers downloading one file, could try to remove the
same lockfile.

This refactor attempts to solve the situation, by using sqlite, since
it is able to handle on it's own database locking and atomic
transactions (With AutoCommit), so once a DDL is issued, sqlite handles
automatically the persistance of those bits.

In case the cache.db file is not present, it will be deployed from
scratch.

Since we are saving now size and etag, we can now easily know if an
asset has been changed, by sending the etag during the first request
and bailing out if there are no changes.

The limiting of assets is done in two stages: at the init of the cache
and after the download of an asset (304 for now are ignored).

Once this is deployed, the cache.db file present in the cache directory
needs to be removed as they are not compatible.