Update cache full stack test
With the new implementation of the cache, fullstack test
needs to be changed.

Mainly this aims to testing the LRU handling on the cache,
for now is simple limit testing.

Download of assets, and check for assets that have been already
downloaded (using the etag to check), is also being tested.