Use cache.sqlite instead of cache.db
To avoid problems with deployment and conflicts with
old implementation, cache.sqlite will be the filename.

Test deployment on dirty cache

In case it is not present, all elements in the cache
location will be removed, and tmp directory will also be
created, as this is where mojolicious will attempt to
download the files.

Add test for assets not found.