Make 'TEST' the only proper mandatory test setting (#1126)
In before there was no real check for most settings implicitly allowing
undefined settings but causing perl warnings. This commit now properly
tests job results where no setting is set besides the name composed of
'TEST' itself. It is not pretty in all cases when the individual
string components are empty but it works.

This commit with the according database schema upgrade makes the values DISTRI
VERSION FLAVOR ARCH BUILD required in the schema with default to empty strings.
Also, there is a database migration script to find all current undef values in
the database and update to empty string.

The in-place job settings migration and transfer on update is updated
accordingly to avoid explicitly overwriting the values with undef in the
database which would null the entries which is not allowed anymore. Instead,
the database takes care to provide a valid empty string as value in the
according cases.

Not fully covered is:
* handle it in UI accordingly where needed e.g. to prevent strings like '--.'
* review the API entry points for additional checks for these settings and
harmonizing them

Related progress issue: https://progress.opensuse.org/issues/6762