Replace die with log_error
To avoid the worker locking itself, the die instruction had to go away.

The problem lies in the try_lock_asset, since most workers will wait
here and timers start at the same time, there will be a race condition
at the point where said function it's called. It could happen also with
update_asset method, but it's most unlikely... But if it fails at this
point, there's nothing much we can do other than let the job be
restarted by the webui