Allow plugins to load specific configurations
With this change we allow the plugins (and not only) to require specific configurations to be loaded.

A Perl module or a Mojolicious Plugin (based on Mojo::Base or Mojolicious::Plugin) if has the method (or attribute, in case of Mojolicious::Plugin) "configuration_fields" and returns a HASH reference, the key/value pairs are cycled to get the relevant configuration from the Ini file parser, and then fill the application config appropriately.

E.g. if we declare a plugin like so:

	package OpenQA::MyPlugin::Foo;
	use Mojo::Base -base;
	has 'configuration_fields' => sub {
	    {
	        auth => {
	            method => 1
	        }};
	};
	1;

The server application will fetch the "auth" section and "method" parameter from the Ini parser object, populating the configuration structure inside the application.

See related Progress issue: https://progress.opensuse.org/issues/6192