Scheduler: Wrap the job_grab query into the worker update transaction
Ensure that the query for searching the most appropriate job to pick is
inside the same transaction that updates the worker state.
While the filtering conditions are built as queries and don't need to be in the transaction,
the search()->first() needs instead to be tied within worker status update.

Also refactor the job_grab loop to try at least a default number of
attempts before bailing out when no jobs are available.
(During tests after ~3 attempts we got a job)

Related Progress Issue: https://progress.opensuse.org/issues/20002