Schedule jobs without polling
The change make the Scheduler DBus service act as a job allocator.
With this change workers don't poll the job_grab api anymore, instead
the scheduler will send the jobs to the workers when they can be
allocated.

To avoid congestions, cpu load peaks and faulty database accesses, the
scheduler can be tweaked to become the bottleneck, and if congestion
control is enabled, it will backoff from scheduling depending on the
load and faults in allocation